+++ R/zzz.R
@@ -27,7 +27,8 @@
     optram.tileid = NA, # allow to filter download by Sentinel tile ID
     optram.scm_mask = TRUE, # mask out clouds, cloud shadow, snow, water.
     optram.overwrite = FALSE, # whether to re-download previously acquired images
-    optram.save_img_list = FALSE # whether to save the list of available images
+    optram.save_img_list = FALSE, # whether to save the list of available images
+    optram.resolution = 10 # What output resolution (10, 20, or 60)
   )
 
   op <- options()